Add PanModalPresentableViewController
